timers internally now use 64bit unsigned int, if available External timer info (glutGet) unchanged. Some work to do to improve Windows timers beyond what it is now...
git-svn-id: http://svn.code.sf.net/p/freeglut/code/trunk/freeglut/freeglut@1110 7f0cb862-5218-0410-a997-914c9d46530a

+++ freeglut/src/Common/freeglut_state.c
@@ -128,7 +128,7 @@ int FGAPIENTRY glutGet( GLenum eWhat )
         return fgState.Initialised;
 
     case GLUT_ELAPSED_TIME:
-        return fgElapsedTime();
+        return (int) fgElapsedTime();
     }
 
     FREEGLUT_EXIT_IF_NOT_INITIALISED ( "glutGet" );
